+++ test/unit/SchemaMigrationTest.php
@@ -169,6 +169,7 @@ class SchemaMigrationTest extends TestCase {
 			'wp_sam_tor_exit_nodes',
 			'wp_sam_asn_cache',
 			'wp_sam_geoip_cache',
+			'wp_sam_detector_policies',
 		);
 
 		$this->assertSame(